% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1.R
\name{m1}
\alias{m1}
\title{Build a url in markdown}
\usage{
m1()
}
\value{
Character.
}
\description{
This function helps users to paste url in markdown
}
\examples{
m1("https://github.com/JiaxiangBU/add2md/")
m1("https://github.com/JiaxiangBU/add2md/","add2md")
}
\author{
Jiaxiang Li
}
