% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3.R
\name{m3}
\alias{m3}
\title{Bold your text in markdown}
\usage{
m3(name = "Type something")
}
\arguments{
\item{name}{Character.}
}
\value{
Character.
}
\description{
This function helps users to paste bold text in markdown
}
\examples{
m3("Jiaxiang Li")
}
\author{
Jiaxiang Li
}
