% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_detail.R
\name{m_detail}
\alias{m_detail}
\title{Build detail and summary html5 style in markdown}
\usage{
m_detail()
}
\arguments{
\item{name}{Character.}
}
\value{
Character.
}
\description{
This function helps users to fold the details by html5 tags.
}
\author{
Jiaxiang Li
}
