% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{m2}
\alias{m2}
\title{Build an image in markdown}
\usage{
m2()
}
\arguments{
\item{url}{Character.}

\item{name}{Character.}
}
\value{
Character.
}
\description{
This function helps users to paste image in markdown
}
\author{
Jiaxiang Li
}
