% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_detail.R
\name{m_detail}
\alias{m_detail}
\title{Bold your text in markdown}
\usage{
m_detail()
}
\arguments{
\item{name}{Character.}
}
\value{
Character.
}
\description{
This function helps users to paste bold text in markdown
}
\author{
Jiaxiang Li
}
